Gamma Completed
Finished implementation of Gamma version of Setback.  All end of game
scenarios are covered now.  Began laying the groundwork of Delta.  Moved
card dealing out of the GameControllers and into the new
CardDealerControllers.  Will need to implement a new
CardDealerController for Delta version, but Alpha, Beta, and Gamma are
currently covered.  Coverage of pre-Delta classes are at 100% except for
Round, which will need a lot of tests crafted specifically for the
purpose of coverage.  Wouldn't be a bad idea to get to that, but some of
it may be covered when Delta testing commences.  The remainder should
get specifically covered after Delta is implemented, as the next goal at
that point will be to develop a GUI and have observables working.