Delta Completed
Fully implemented the Delta version of Setback.  This includes randomly
dealt hands.  Randomness or predictability is determined at the factory
level, based on whether or not a seed is provided when creating a
DeltaSetbackGame.  Discarding of cards is also handled in this version.
Versions prior to Delta do not require (or allow for) cards to be
discarded at the start of a trick.  From this point onward each player
must discard three valid cards.  The players can discard in any order.