Networking
Created a SetbackServer, which is a multithreaded server that spawns new
SetbackServerThreads every time a SetbackClient connects to the server.
Currently the server runs on the localhost at port 2323, but this can be
overwritten on the command line.  Additionally, created a Command enum
that has the different commands that a player may need.  A
CommandMessage contains this Command plus any additional arguments that
may be needed.  This is parsed by the CommandParser class, which is
covered by tests.  The PlayerController is the class that a player will
actually interact with.  Later it will have a GUI component, but for now
it is text based.  The next step is implementing the SetbackObserver and
SetbackObservable classes.