Communication Update
Updated the ServerThread and Client to be able to communicate via
updates.  When an update is sent out by the SetbackGameController, that
message is put in the print queue for each client.  It is printed out
the next time the client hits a newline.  Also made it so a single
client cannot select all four players.