SortHand and GUI Magic
Created the SortHand method in the Hand class which sorts the cards in
the hand from least to greatest.  Also worked on the GUI, which now has
a PlayerSelectView.  Need to add rejection messages to that screen.  The
next screen is the PleaseWaitView.  Need to update the message in there.
Finally is the PlaceBetsView which will need to display the player's
hand, as well as any bets made by other players.