Live Updating
Figured out how to get the Client to periodically check for updates from
the server using a Timer.  As a result the PleaseWaitView is no longer
needed.  Instead, the PlaceBetsView defaults to showing the please wait
message, and checks to see if everyone is connected (by looking for its
hand) every two seconds.