Betting GUI
Betting now works up to the point of placing the fourth bet.  Goals for
next time: Have a card back in a file called Back.png which will allow
it to be displayed in the GUI, and working on resolving bets.  Also need
a way to tell the user that their bet was invalid.  Or just screw the
user and assume they know the rules.