CardBack and SelectTrumpView
Created an image for the back of a Card.  Also, the SelectTrumpView is
working.  Might have missed one or two things in it, but overall it gets
the job done.  Next screen to do is discarding.  Testing is becoming
bothersome since the screens I need to test are several layers down.
Such is life.