Refresh Works and Server is Fancy
Got refreshing to work on my laptop.  Looks like the issue was just that
the GUI was too large for the itty bitty screen.  As a result, many of
the GUIs need to be fixed to handle the smaller size.  These tasks are
outlined in the TODO.  Also, made the Server switch between retail and
debug based on environment variables.