PlayerSelectView Testing
The PlayerSelectView class is now automatically tested.  Only downside
is that each test pops open a client view for a split second, making it
look fairly stupid when it runs the tests.  Also added an error message
to the screen if you try to select a player that has already been taken.