PleaseWaitView and PlaceBetsView
Created a PleaseWaitView that handles waiting for all of the players to
connect.  Previously, this functionality was hidden in the
PlaceBetsView, but it has been separated for ease of testing, and better
coding practices.  Also moved the SetbackClientView class from the
client package to the view package to make imports nicer, and testing
easier.  Coverage is climbing back up to a more respectable figure now
that the GUI is getting covered.  The next class to cover is the
TrumpSelectionView.  Make sure everything is covered and all of the
warnings and notification texts are visible in the GUIs.